% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbt.R
\name{compile_analyses}
\alias{compile_analyses}
\title{Compile the SQL of a dbt Analysis}
\usage{
compile_analyses(analysis_name, dbt_conn, ...)
}
\arguments{
\item{analysis_name}{Name of the analysis (without .sql file ending)}

\item{dbt_conn}{dbt connection}

\item{...}{Variables to pass to dbt compile}
}
\value{
The compiled SQL statement
}
\description{
Compile the SQL of a dbt Analysis
}
