% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbt.R
\name{dbt_conn}
\alias{dbt_conn}
\title{Connection to dbt Project}
\usage{
dbt_conn(project_path, dbt_path = NULL)
}
\arguments{
\item{project_path}{Path to dbt project}

\item{dbt_path}{Path to dbt executable (default is {dbt_project}/venv/bin/dbt)}
}
\value{
a dbt connection
}
\description{
Connection to dbt Project
}
