% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_utils.R
\name{drop_all_tables_in_schema}
\alias{drop_all_tables_in_schema}
\title{Drop All Tables in a Postgres Schema}
\usage{
drop_all_tables_in_schema(schema_name)
}
\arguments{
\item{schema_name}{}
}
\description{
Drop All Tables in a Postgres Schema
}
