% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbt.R
\name{get_project_name}
\alias{get_project_name}
\title{Get dbt Project Name}
\usage{
get_project_name(dbt_conn)
}
\arguments{
\item{dbt_conn}{dbt connection}
}
\value{
The project name (character)
}
\description{
Get dbt Project Name
}
